pyproject.toml
[project]
name = "friday-tony-stark-demo"
version = "0.1.0"
description = "Friday — A Tony Stark-inspired MCP server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp",
    "httpx",
    "livekit-agents[deepgram,groq,openai,sarvam,silero]>=1.5.1",
    "livekit-plugins-google>=1.5.1",
    "python-dotenv",
    "beautifulsoup4",
    "html2text",
    "gtts",
    "pydub",
    "cloudflare",
    "groq>=1.1.2",
    "starlette>=1.0.0",
]

[project.scripts]
friday = "server:main"
friday_voice = "agent_friday:dev"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server.py", "friday"]
